pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-data-cleaning-env"
version = "0.1.0"
description = "Data Cleaning Environment for OpenEnv — AI agents learn to clean messy tabular data"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv dependencies
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    # Data handling
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
inference = [
    "openai>=1.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["data_cleaning_env", "data_cleaning_env.server"]
package-dir = { "data_cleaning_env" = ".", "data_cleaning_env.server" = "server" }
